fix: bad registration results
Due to the downloaded thumb, instead of the whole image, the
coordinates of the annotations are not properly deformed. To fix
this, annotations are first mapped from the Cytomine image, to the
local file, then deformed using VALIS, then mapped to the Cytomine
image of destination.
This commit also fixes a few issues here and there.