fix: wrong argument filled in by server
(Possibly) due to the common prefix, when max_proc_size_micro
is not set, but max_proc_size is set to e.g. 800, the former gets
filled in as "800_micro" which is not a valid integer. Hopefully,
this fixes this issue by avoiding the common prefix.

Also: small cleanups to make the code more readable